Organize Notification related components
