+++ src/components/Navigation/Notifications/NotificationReply.js
@@ -1,0 +1,39 @@
+import React, { PropTypes } from 'react';
+import classNames from 'classnames';
+import { FormattedRelative } from 'react-intl';
+import { Link } from 'react-router';
+import Avatar from '../../Avatar';
+
+const NotificationReply = ({ onClick, id, read, date, payload }) =>
+  <div
+    onClick={() => onClick(id)}
+    className={
+      classNames('Notifications__notification', {
+        'Notifications__notification--unread': !read,
+      })
+    }
+  >
+    <Avatar username="guest123" size={40} />
+    <div className="Notifications__notification__text">
+      <div className="Notifications__notification__text__message">
+        <Link to={`/${payload.user}`}>{payload.user}</Link> replied on your <Link to={payload.post_url}>post</Link>.
+      </div>
+      <div className="Notifications__notification__text__date">
+        <FormattedRelative value={date} />
+      </div>
+    </div>
+  </div>;
+
+NotificationReply.propTypes = {
+  onClick: PropTypes.func,
+  id: PropTypes.number,
+  read: PropTypes.bool,
+  date: PropTypes.string,
+  payload: PropTypes.shape(),
+};
+
+NotificationReply.defaultProps = {
+  onClick: () => {},
+};
+
+export default NotificationReply;
