+++ src/components/Navigation/Notifications/Notifications.less
@@ -1,4 +1,4 @@
-@import "../../styles/custom.less";
+@import "../../../styles/custom.less";
 
 .Notifications {
   margin: -8px -16px;
